Move the `OwnerReferencesPermissionEnforcement` admission plugin to the `k8s.io/apiserver` module
In particular, this commit is build by executing:
```
mv plugin/pkg/admission/gc staging/src/k8s.io/apiserver/pkg/admission/plugin
```

Then, the import to the admission plugin in `pkg/kubeapiserver/options/plugins.go` is adapted.
